WIP: Add coreos-inject-rootmap.service, unlock root-on-RAID
This implements the rootmap code that figures out all the dependencies
required to find `/sysroot`, and injects them into the BLS config. For
more information, see:

https://github.com/coreos/fedora-coreos-tracker/issues/94#issuecomment-650433166

This supports RAID and LUKS devices, though the latter needs outstanding
work in Ignition to actually be fully supported.

This also implements the `root=UUID=$uuid` inject patch proposed in
https://github.com/coreos/fedora-coreos-tracker/issues/465.

With this, it's now possible to configure FCOS with root on a RAID
device!

Closes: #465