ignition-ostree-growfs: fix dependency to sysroot mount
In the case of multipath,
`ignition-ostree-mount-firstboot-sysroot.service` won't kick in because
the `root=` karg is already present on first boot. But we still need to
grow the partition and filesystem in that case.

What we really mean here is that the sysroot should already be mounted
before we growfs. Drop the `Requires=` and add `sysroot.mount` to the
list of `After=` so that we cover the case where the sysroot is mounted
via the mount unit generated via systemd-fstab-generator from the
`root=` karg.

(cherry picked from commit caa586a5f7dc9db58624aee12c24755f7dc0ac85)