ignition-ostree-uuid-root: also run if root= is provided
We want it to run in the multipath case, in which users may provide a
`root=` karg. So drop that condition.

This service is already queued against `ignition-diskful.target` so it
only runs on first boot. It also already correctly does nothing if
somehow the UUID was already changed.

(cherry picked from commit 5b96230305b3a1784daa6e94729e06e5509c74f6)