05core: run `coreos-gpt-setup` and `ignition-ostree-uuid-boot` later
These services need to manipulate the boot disk but if the disk is
backed by iSCSI, we need networking first. Move them to run later in the
boot process but still before `ignition-disks`.

We still need `coreos-copy-firstboot-network.service` to depend on the
bootfs being available though. Do that now by explicitly adding the deps
like we do with `coreos-ignition-setup-user.service`.