05core: adapt `coreos-ignition-setup-user.service` for iSCSI boot
In an iSCSI boot, we need networking before we can see the bootfs. So
`coreos-ignition-setup-user.service` needs to run after networking. We
also can't rely on the `fetch-offline` stage. Tweak things so that in
an iSCSI boot (1) we neuter the `fetch-offline` stage, and (2) we still
ensure that `coreos-ignition-setup-user.service` runs before the `fetch`
stage, which will take care of "consuming" the injected Ignition config.