manifests: just disable systemd-firstboot, don't nuke
In the spirit of abusing `remove-from-packages` less and being more
closely aligned with tier-x, stop literally nuking the systemd-firstboot
binary and systemd unit. Instead, just disable the systemd unit.

Note it's also disabled in the presets (in `40-coreos-systemd.preset`)
and as mentioned in the comment, via a runtime dropin. So that's triple
disablement already.